[GPU] Strided slice primitive params update (#11339)

+++ src/plugins/intel_gpu/include/intel_gpu/primitives/strided_slice.hpp
@@ -38,10 +38,10 @@ struct strided_slice : public primitive_base<strided_slice> {
                   const primitive_id& begin_id,
                   const primitive_id& end_id,
                   const primitive_id& strides_id,
-                  std::vector<uint8_t> begin_mask,
-                  std::vector<uint8_t> end_mask,
-                  std::vector<uint8_t> new_axis_mask,
-                  std::vector<uint8_t> shrink_axis_mask,
+                  std::vector<int64_t> begin_mask,
+                  std::vector<int64_t> end_mask,
+                  std::vector<int64_t> new_axis_mask,
+                  std::vector<int64_t> shrink_axis_mask,
                   const ov::Shape out_size,
                   const primitive_id& ext_prim_id = "",
                   const padding& output_padding = padding())
@@ -53,13 +53,13 @@ struct strided_slice : public primitive_base<strided_slice> {
           out_size(out_size) {}
 
     /// @brief Array of bits, that provide replace begin[i] to max possible range in that dimension.
-    std::vector<uint8_t> begin_mask;
+    std::vector<int64_t> begin_mask;
     /// @brief Array of bits, that provide replace end[i] to max possible range in that dimension.
-    std::vector<uint8_t> end_mask;
+    std::vector<int64_t> end_mask;
     /// @brief Array of bits, that provide adding a new length 1 dimension at ith position in the output tensor.
-    std::vector<uint8_t> new_axis_mask;
+    std::vector<int64_t> new_axis_mask;
     /// @brief Array of bits, that provide shrinks the dimensionality by 1, taking on the value at index begin[i].
-    std::vector<uint8_t> shrink_axis_mask;
+    std::vector<int64_t> shrink_axis_mask;
     /// @brief Size of output tensor
     ov::Shape out_size;
 };
